# Fuse q_a_proj and kv_a_proj for DeepSeek models