<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## Motivation

In self attention module of DeepSeek v3, the `q_a_proj` and `kv_a_proj` both takes hidden state as input, so they can be fused into one module, and one launching of DeepGemm is saved.

<!-- Explain the purpose of this PR and the goals it aims to achieve. -->

## Modifications

<!-- Describe the changes made in this PR. -->

When `q_lora_rank` is more than 0 (which is true for DeepSeek V3 and R1), `self.q_a_proj` and `self.kv_a_proj_with_mqa` will be fused into a new module `self.fused_qkv_a_proj_with_mqa`. When loading weights, the weights and block scales of `q_a_proj` and `kv_a_proj` will be concatenated and loaded into `self.fused_qkv_a_proj_with_mqa`.

## Accuracy Test

### Launch
```bash
SGL_ENABLE_JIT_DEEPGEMM=1 python3 -m sglang.launch_server --model /dev/shm/DeepSeek-V3 --tp 8 --trust-remote-code
```

### GSM8K
```bash
python3 benchmark/gsm8k/bench_sglang.py --num-shots 8 --num-questions 1319 --parallel 1319
```

```bash
Accuracy: 0.951
Invalid: 0.000
Latency: 73.598 s
Output throughput: 1949.868 token/s
```

### MMLU

```bash
bash benchmark/mmlu/download_data.sh
python3 benchmark/mmlu/bench_sglang.py --nsub 100 --ntrain 5 --parallel 2000
```

```bash
Total latency: 110.354
Average accuracy: 0.875
```

### Benchmark

```bash
SGL_ENABLE_JIT_DEEPGEMM=1 python3 -m sglang.launch_server --model /dev/shm/DeepSeek-V3 --tp 8 --trust-remote-code --attention-backend fa3 --disable-radix

python3 -m sglang.bench_one_batch_server --model None --base-url http://0.0.0.0:30000 --batch-size 1 --input-len 512 --output-len 512
```


Main branch:
```bash
batch size: 1
latency: 10.85 s
output throughput: 47.21 token/s
(input + output) throughput: 94.41 token/s

batch size: 32
latency: 16.83 s
output throughput: 973.37 token/s
(input + output) throughput: 1946.74 token/s
```

This PR:
```bash
batch size: 1
latency: 10.41 s
output throughput: 49.16 token/s
(input + output) throughput: 98.33 token/s

batch size: 32
latency: 16.21 s
output throughput: 1010.55 token/s
(input + output) throughput: 2021.11 token/s
```

About 4.2% improvement for bs=1 and 3.8% for bs=32


### Profile

```bash
SGL_ENABLE_JIT_DEEPGEMM=1 python3 -m sglang.bench_one_batch --model-path /dev/shm/DeepSeek-V3 --tp 8 --cuda-graph-max-bs 32 --batch 16 --input-len 1024  --output-len 2 --profile
```

Main Branch: 4 DeepGemm kernels before FlashAttention
![截屏2025-04-22 16 11 26](https://github.com/user-attachments/assets/39e5510f-126f-4041-a609-9b56c2e5b672)

This PR: 3 DeepGemm kernels before FlashAttention
![截屏2025-04-22 16 12 08](https://github.com/user-attachments/assets/083caed2-30a4-4675-99c9-ecbc1340cd59)



## Checklist

- [ ] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).
- [ ] Add unit tests as outlined in the [Running Unit Tests](https://docs.sglang.ai/references/contribution_guide.html#running-unit-tests-adding-to-ci).
- [ ] Update documentation / docstrings / example tutorials as needed, according to [Writing Documentation](https://docs.sglang.ai/references/contribution_guide.html#writing-documentation-running-docs-ci).
- [ ] Provide throughput / latency benchmark results and accuracy evaluation results as needed, according to [Benchmark and Profiling](https://docs.sglang.ai/references/benchmark_and_profiling.html) and [Accuracy Results](https://docs.sglang.ai/references/accuracy_evaluation.html).
- [ ] For reviewers: If you haven't made any contributions to this PR and are only assisting with merging the main branch, please remove yourself as a co-author when merging the PR.
- [ ] Please feel free to join our Slack channel at https://slack.sglang.ai to discuss your PR.